7192416_APIs_DRF-main/requirements.txt
djangorestframework
djangorestframework-simplejwt

Django~=4.1.7